backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
boto3==1.28.85
Pillow==10.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
aws-xray-sdk==2.12.1
PyJWT==2.8.0
bcrypt==4.0.1